Tehtävä 2.13
