+++ osa2/maiden-tiedot/src/components/NaviButton.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+//Navigointinappulaelementti
+const NaviButton = ({name, handleButtonClick}) => {
+    console.log('Piirretään napula', name, 'funktio:', handleButtonClick)
+    return(
+      <button onClick={() =>handleButtonClick({name})}>show</button>
+    )
+}
+
+export default NaviButton
